* Fix a problem with compilation of libffi that I introduced in my previous checkin * Explictly add '-fexceptions' to the compiler arguments (probably not needed) * Fix some 64-bit issues (NSInteger vs. int issues) * More serious: code sometimes looked at PyObjCObject fields without checking that   the object really is a PyObjCObject, even when objects can PyObjCClass instances.   This happens to work fine on all platforms but PPC64, but that's accidental. * Enable '@try{' instead of 'NSDURING'
All of this is part of trying to get pyobjc-core going on a PPC64 system. Sadly
enough I now get a fatal error from the ObjC runtime because of an uncaught
exception, even though code is passing through an @try block that should catch
the exception.

+++ pyobjc-core/Modules/objc/OC_PythonData.m
@@ -63,7 +63,7 @@
 			PyErr_Clear();
 			rval = 0;
 		}
-		if ((unsigned)buffer_len > NSUIntegerMax) {
+		if ((NSUInteger)buffer_len > NSUIntegerMax) {
 			rval = NSUIntegerMax;
 		} else {
 			rval = buffer_len;
